semi-large changes to parse and ast
reduce stratification of Parse, which lets us write recursive parser
bits without open recursion and fixpoints. Required because we have
mutually recursive ones now (assignment and expression) and doing that
with fixpoints is way more annoying that it is worth. We can still
restrict the interface of the entire parse at the file scope or at the
parent module and it should be fine.

This is all for setting up modules and top level assignments